[#178] Moved Data Structures
The data structures (Graph, LinkedList) have been moved into the DataStructures folder
