+++ app/src/main/java/DataStructures/Graph.java
@@ -1,4 +1,4 @@
-package com.example.scoutconcordia;
+package DataStructures;
 
 import com.google.android.gms.maps.model.LatLng;
