+++ app/src/main/java/DataStructures/LinkedList.java
@@ -1,14 +1,6 @@
-package com.example.scoutconcordia;
+package DataStructures;
 
-import java.io.Serializable;
 import java.io.File;
-import java.io.ObjectInputStream;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.ObjectOutputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.util.Collection;
 
 public class LinkedList <E>
 {
